Fixed the pyright errors
