+++ examples/simulation.py
@@ -123,7 +123,7 @@ def get_argparser() -> argparse.ArgumentParser:
 
 
 if __name__ == "__main__":
-    # Initialize the configuration and apply overrides from the command line arguments.
+    # Instantiate the config using the command line arguments as overrides.
     args = get_argparser().parse_args()
     override_dict = {}
     if args.trading_days is not None:
@@ -132,23 +132,21 @@ if __name__ == "__main__":
         override_dict["num_blocks_per_day"] = args.blocks_per_day
     if args.log_level is not None:
         override_dict["logging_level"] = args.log_level
+    override_dict["pricing_model_name"] = args.pricing_model
     override_dict["vault_apr"] = {"type": "GeometricBrownianMotion", "initial": 0.05}
-    config_ = config_utils.override_config_variables(
-        config_utils.load_and_parse_config_file(args.config), override_dict
-    )
+    config = config_utils.override_config_variables(config_utils.load_and_parse_config_file(args.config), override_dict)
 
     # Define root logging parameters.
     output_utils.setup_logging(
         log_filename=args.output,
         max_bytes=args.max_bytes,
-        log_level=config_utils.text_to_logging_level(config_.simulator.logging_level),
+        log_level=config_utils.text_to_logging_level(config.simulator.logging_level),
     )
 
     # Initialize the simulator.
-    rng = np.random.default_rng(config_.simulator.random_seed)
-    pricing_model = sim_utils.get_pricing_model(model_name=args.pricing_model)
+    rng = np.random.default_rng(config.simulator.random_seed)
     agents = get_example_agents(rng=rng, new_agents=args.num_agents, existing_agents=1)
-    simulator = sim_utils.get_simulator(config_, rng, pricing_model, agents)
+    simulator = sim_utils.get_simulator(config, rng, agents)
 
     # Run the simulation.
     simulator.run_simulation()
