+++ src/elfpy/simulators.py
@@ -1,24 +1,102 @@
 """Simulator class wraps the pricing models and markets for experiment tracking and execution"""
 
 from __future__ import annotations  # types will be strings by default in 3.11
-from typing import Optional, TYPE_CHECKING
+from typing import Callable, Optional, TYPE_CHECKING
 import datetime
 import json
 import logging
-
 import numpy as np
 from numpy.random._generator import Generator
+from stochastic.processes import GeometricBrownianMotion
 
-from elfpy.utils import sim_utils  # utilities for setting up a simulation
 import elfpy.utils.time as time_utils
 from elfpy.utils.outputs import CustomEncoder
-from elfpy.types import MarketDeltas, SimulationState
+from elfpy.types import MarketDeltas, RandomSimulationVariables, SimulationState
 
 if TYPE_CHECKING:
     from elfpy.agent import Agent
     from elfpy.markets import Market
-    from elfpy.pricing_models.base import PricingModel
     from elfpy.utils.config import Config
+
+
+def setup_vault_apr(config, rng):
+    """Construct the vault_apr list
+    Note: callable type option would allow for infinite num_trading_days after small modifications
+
+    Arguments
+    ---------
+    config : Config
+        config object, as defined in elfpy.utils.config
+    rng : Generator
+        random number generator; output of np.random.default_rng(seed)
+
+    Returns
+    -------
+    vault_apr : list
+        list of apr values that is the same length as num_trading_days
+    """
+    if isinstance(config.market.vault_apr, dict):  # dictionary specifies parameters for the callable
+        match config.market.vault_apr["type"].lower():
+            case "constant":
+                vault_apr = [
+                    config.market.vault_apr["value"],
+                ] * config.simulator.num_trading_days
+            case "uniform":
+                vault_apr = rng.uniform(
+                    low=config.market.vault_apr["low"],
+                    high=config.market.vault_apr["high"],
+                    size=config.simulator.num_trading_days,
+                ).tolist()
+            case "geometricbrownianmotion":
+                # the n argument is number of steps, so the number of points is n+1
+                vault_apr = (
+                    GeometricBrownianMotion(rng=rng)
+                    .sample(n=config.simulator.num_trading_days - 1, initial=config.market.vault_apr["initial"])
+                    .tolist()
+                )
+            case _:
+                raise ValueError(
+                    f"{config.market.vault_apr['type']=} not one of \"constant\","
+                    f'"uniform", or "geometricbrownianmotion"'
+                )
+    elif isinstance(config.market.vault_apr, Callable):  # callable (optionally generator) function
+        vault_apr = [config.market.vault_apr() for _ in range(config.simulator.num_trading_days)]
+    elif isinstance(config.market.vault_apr, list):  # user-defined list of values
+        vault_apr = config.market.vault_apr
+    elif isinstance(config.market.vault_apr, float):  # single constant value to be cast to a float
+        vault_apr = [float(config.market.vault_apr)] * config.simulator.num_trading_days
+    else:
+        raise TypeError(
+            f"config.market.vault_apr must be an int, list, dict, or callable, not {type(config.market.vault_apr)}"
+        )
+    return vault_apr
+
+
+def get_random_variables(config, rng):
+    """Use random number generator to assign initial simulation parameter values
+
+    Arguments
+    ---------
+    config : Config
+        config object, as defined in elfpy.utils.config
+    rng : Generator
+        random number generator; output of np.random.default_rng(seed)
+
+    Returns
+    -------
+    RandomSimulationVariables
+        dataclass that contains variables for initiating and running simulations
+    """
+    random_vars = RandomSimulationVariables(
+        target_liquidity=rng.uniform(low=config.market.min_target_liquidity, high=config.market.max_target_liquidity),
+        target_pool_apr=rng.uniform(
+            low=config.amm.min_pool_apr, high=config.amm.max_pool_apr
+        ),  # starting fixed apr as a decimal
+        fee_percent=rng.uniform(low=config.amm.min_fee, high=config.amm.max_fee),
+        vault_apr=setup_vault_apr(config, rng),
+        init_vault_age=rng.uniform(low=config.market.min_vault_age, high=config.market.max_vault_age),
+    )
+    return random_vars
 
 
 class Simulator:
@@ -36,9 +114,8 @@ class Simulator:
         self,
         config: Config,
         market: Market,
-        agents: dict[int, Agent],
         rng: Generator,
-        random_simulation_variables: Optional[sim_utils.RandomSimulationVariables] = None,
+        random_simulation_variables: Optional[RandomSimulationVariables] = None,
     ):
         # pylint: disable=too-many-arguments
         # pylint: disable=too-many-statements
@@ -49,11 +126,11 @@ class Simulator:
         self.market = market
         self.set_rng(rng)
         if random_simulation_variables is None:
-            self.random_variables = sim_utils.get_random_variables(self.config, self.rng)
+            self.random_variables = get_random_variables(self.config, self.rng)
         else:
             self.random_variables = random_simulation_variables
         self.check_vault_apr()
-        self.agents = agents
+        self.agents = {}
 
         # Simulation variables
         self.run_number = 0
