+++ tests/test_logging.py
@@ -14,7 +14,6 @@ from typing import Any
 import numpy as np
 
 from elfpy.utils.parse_config import load_and_parse_config_file
-from elfpy.simulators import Simulator
 from elfpy.utils import sim_utils, outputs as output_utils  # utilities for setting up a simulation
 import elfpy.utils.parse_config as config_utils
 
@@ -28,8 +27,7 @@ class BaseLogTest(unittest.TestCase):
         # Initialize the simulator.
         config = config_utils.override_config_variables(load_and_parse_config_file(config_file), override_dict)
         rng = np.random.default_rng(config.simulator.random_seed)
-        pricing_model = sim_utils.get_pricing_model(model_name=config.amm.pricing_model_name)
-        simulator = sim_utils.get_simulator(config, rng, pricing_model, [])
+        simulator = sim_utils.get_simulator(config, rng)
 
         # Run the simulation.
         simulator.run_simulation()
